Fix Sonos track metadata parsing failure (Spotify session)
Aparently, if there's a Spotify session active in the Sonos unit
`get_current_track_info` yields something akin to:

  track_info = {
    'title': '',
    'artist': '',
    'album': '',
    'album_art': '',
    'position': 'NOT_IMPLEMENTED',
    'playlist_position': '1',
    'duration': 'NOT_IMPLEMENTED',
    'uri': 'x-sonos-vli:RINCON_542A1B61190801400:2,spotify:3ce01b54b0c03bd9a42869a19f566f92',
    'metadata': 'NOT_IMPLEMENTED'
  }

Trying to parse that metadata field as an XML document was resulting in
a ParseError.

However, album selection in songbird was not working at all and I had to
hop into the Sonos app to click 'End session'. I'm guessing the *correct*
way to go about this is would be to terminate third-party sessions on an
album selection.
I see that the `soco.core.SoCo` class has an
`end_direct_control_session` method -- looks promising.