Terminate 3rd-party sessions on app startup
I think this fixes the problem with people leaving the Sonos unit with a
Spotify session active.
An alternative was to make this function call on album selections to
make sure the app can keep working if someone starts another session. I
guess the current solution is just to restart it. \shrug